# openapi_doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WAAPI_mugh",
    "description": "API WhatsApp using bailey",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "SystemApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-system-api-key"
      }
    }
  },
  "paths": {
    "/sessions/{sessionId}/start": {
      "post": {
        "summary": "Start a new socket for a given session ID. Socket Will disconnect after 5 minutes When no QR scan. You should post a new start socket",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The ID of the session to start",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully started the session"
          },
          "500": {
            "description": "Failed to start socket for session"
          }
        }
      }
    },
    "/sessions/{sessionId}": {
      "delete": {
        "summary": "Delete a session by session ID",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The ID of the session to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the session",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Session {sessionId} deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Session {sessionId} not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to delete the session file",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to delete session file"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{sessionId}/status": {
      "get": {
        "summary": "Check if a socket is already started for a given session ID",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The ID of the session",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session is running"
          },
          "404": {
            "description": "Session is not running"
          }
        }
      }
    },
    "/sessions/{sessionId}/qrcode": {
      "get": {
        "summary": "Get the QR code for a specific session ID",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The ID of the session",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved QR code",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "qrCodeUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "QR code not available or socket not running"
          }
        }
      }
    },
    "/sessions/{sessionId}/check/{phone}": {
      "get": {
        "summary": "Check if a number exists on WhatsApp",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The ID of the session",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "in": "path",
            "required": true,
            "description": "The phone number to check",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Phone number exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exists": {
                      "type": "boolean"
                    },
                    "jid": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to check WhatsApp number"
          }
        }
      }
    },
    "/send/{sessionId}/messages": {
      "post": {
        "summary": "Send a message. Phone number include country code without +",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The ID of the session",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "number": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message sent successfully"
          },
          "404": {
            "description": "Session not found or number not registered"
          },
          "500": {
            "description": "Failed to send message"
          }
        }
      }
    },
    "/send/{sessionId}/messages/image": {
      "post": {
        "summary": "Send an image message. Phone number include country code without +",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The ID of the session",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "number": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  },
                  "attachment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image sent successfully"
          },
          "404": {
            "description": "Session not found or number not registered"
          },
          "500": {
            "description": "Failed to send image"
          }
        }
      }
    },
    "/send/{sessionId}/messages/file": {
      "post": {
        "summary": "Send a PDF or document file. Phone number include country code without +",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The ID of the session",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "number": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  },
                  "attachment": {
                    "type": "object",
                    "properties": {
                      "url": {
                        "type": "string"
                      },
                      "fileName": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File sent successfully"
          },
          "404": {
            "description": "Session not found or number not registered"
          },
          "500": {
            "description": "Failed to send attachment"
          }
        }
      }
    },
    "/webhook/{sessionId}": {
      "post": {
        "summary": "Set the webhook URL for a specific session ID",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The ID of the session",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "webhookUrl": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook URL set successfully"
          },
          "400": {
            "description": "Invalid webhook URL"
          }
        }
      },
      "get": {
        "summary": "Get the webhook URL for a specific session ID",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The ID of the session",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved webhook URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": {
                      "type": "string"
                    },
                    "webhookUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Webhook URL not found for session"
          }
        }
      }
    },
    "/key/{sessionId}": {
      "post": {
        "summary": "Generate a new API key for a specific session ID",
        "security": [
          {
            "SystemApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The ID of the session",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": {
                      "type": "string"
                    },
                    "apiKey": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access allowed only from localhost"
          }
        }
      },
      "get": {
        "summary": "Get the API key for a specific session ID",
        "security": [
          {
            "SystemApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The ID of the session",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved API key",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKey": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "API key not found for session"
          }
        }
      },
      "delete": {
        "summary": "Delete the API key for a specific session ID",
        "security": [
          {
            "SystemApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The ID of the session",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key deleted successfully"
          },
          "404": {
            "description": "API key not found for session"
          }
        }
      }
    }
  }
}